refactor(webapp): move activity pages from /app/activities/* to /app/* mirroring mobile routes
- git mv all create/edit page.tsx and page.ui.spec.tsx files
- Update Cancel/Save/Delete redirects from '/app/activities' to '/app'
- Update usePathname mocks in test files to reflect new routes
- Remove unused @ts-expect-error directives
- Fix Id<'activities'> type errors by importing Id from dataModel
- Fix @ts-expect-error in activities/page.tsx (now unused)
