+++ apps/webapp/src/app/app/diaper-change/create/page.tsx
@@ -30,7 +30,6 @@ export default function DiaperCreatePage() {
   const authState = useAuthState();
   const isAuthenticated = authState?.state === 'authenticated';
 
-  // @ts-expect-error — api.web.babyTracker may not be in generated types yet
   const createActivity = useSessionMutation(api.web.babyTracker.activities.create);
 
   const [diaperType, setDiaperType] = useState<DiaperType>('wet');
@@ -55,7 +54,7 @@ export default function DiaperCreatePage() {
         },
       } as any);
 
-      router.push('/app/activities');
+      router.push('/app');
     } finally {
       setSaving(false);
     }
@@ -103,7 +102,7 @@ export default function DiaperCreatePage() {
 
       {/* Actions */}
       <div className="flex gap-3 mt-6">
-        <Button variant="outline" onClick={() => router.push('/app/activities')}>
+        <Button variant="outline" onClick={() => router.push('/app')}>
           Cancel
         </Button>
         <Button onClick={handleSave} disabled={saving}>
