+++ apps/webapp/src/app/app/feed/edit/[activityId]/page.tsx
@@ -4,6 +4,7 @@ import { useRouter, useParams } from 'next/navigation';
 import { useState, useEffect } from 'react';
 import { useSessionQuery, useSessionMutation } from 'convex-helpers/react/sessions';
 import { api } from '@workspace/backend/convex/_generated/api';
+import { Id } from '@workspace/backend/convex/_generated/dataModel';
 
 import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
@@ -35,16 +36,13 @@ export default function FeedEditPage() {
   const authState = useAuthState();
   const isAuthenticated = authState?.state === 'authenticated';
 
-  const activityId = params?.activityId as string;
+  const activityId = params?.activityId as Id<'activities'>;
 
-  // @ts-expect-error — api.web.babyTracker may not be in generated types yet
   const activity = useSessionQuery(api.web.babyTracker.activities.getById, {
     activityId,
   }) as Record<string, unknown> | undefined;
 
-  // @ts-expect-error
   const updateActivity = useSessionMutation(api.web.babyTracker.activities.update);
-  // @ts-expect-error
   const deleteActivity = useSessionMutation(api.web.babyTracker.activities.deleteActivity);
 
   // Form state
@@ -166,7 +164,7 @@ export default function FeedEditPage() {
         },
       } as any);
 
-      router.push('/app/activities');
+      router.push('/app');
     } finally {
       setSaving(false);
     }
@@ -176,7 +174,7 @@ export default function FeedEditPage() {
     setDeleting(true);
     try {
       await deleteActivity({ activityId } as any);
-      router.push('/app/activities');
+      router.push('/app');
     } finally {
       setDeleting(false);
     }
@@ -320,7 +318,7 @@ export default function FeedEditPage() {
 
       {/* Actions */}
       <div className="flex gap-3 mt-6">
-        <Button variant="outline" onClick={() => router.push('/app/activities')}>
+        <Button variant="outline" onClick={() => router.push('/app')}>
           Cancel
         </Button>
         <Button onClick={handleSave} disabled={saving}>
